requirements.txt
numpy
pandas
pillow
scikit-image
scikit-learn
matplotlib
jupyterlab
opencv-python
torch
torchvision

